Cleanup LibrarySearch.get_search_items
This also caught a missed migration of library_return_search_items.

There is a functional change in how get_search_items works, in that it now
accumulates the playtime, num_songs, and results over multiple searches it
finds in LibrarySearch.get_searches.

I expect this is how it was meant to behave since it got changed back in
813140c6.  The other similar functions in LibrarySearch, get_count and
get_list_items both follow this model.  But I could be wrong.  If so, it
should operate on searches[-1].