Revamp Artwork/Library divide (and threading)
This is an Intermediate Commit.  More cleanup and tweaking to follow.

So far:
  1. Explicitly split off an ArtworkThread class for handling the thread acts
  2. It talks to consumers via a custom GTK+ signal ('art_ready')

The current signal consumer is the library view, but in subsequent commits the
breadcrumb and other artwork consumers can be made to respond appropriately.

Currently there is a bit more lag than desired for filling the library view.
There will likely always be a tradeoff of responsiveness for this, but it can
be faster.