LibrarySearch, LibraryView methods use SongRecords
In order to simplify the code a bit, we can have both LibrarySearch
and LibraryView methods accept SongRecord objects instead of the fields
from those objects.

This makes it a bit simpler, because we are passing along single objects
instead of some more complex set of arguments that are just unrolled
SongRecord objects.