Rework Library's filter search
Major change, includes:
1. Split out the thread into its own class
2. Pull Search-related functionality into LibrarySearch

Note that before this change the threading was useless (AFAICT), in that it
would simply call GLib.idle_add to perform the actual searching, and that
runs on the main thread.

The actual calls to MPD still happen in the main thread, due to stability
issues if they happen in the child thread.  But the actual work of filtering
the results is now done in the child thread.

Given that nobody apparently complained that the work happened on the main
thread before, it may be preferrable to remove the threading entirely, but
in my testing this seems to work just as well.  The only big tradeoff seems
to be that of code complexity.