current: progressively add songs to Current instead of all at once
This prevents Sonata from freezing while updating the Current's model if
there are lot of songs currently selected.

There are probably tons of improvements to do, such as adding a progress
bar to know how much Sonata did and how much left to do, and stopping the
loading if things are changing before it finished updating the model.

This should help #47