current: format the current playlist values as late as possible
This is a first commit trying to improve the speed of the Current playlist.

This part actually simplifies the Current model a lot, since it only stores
a song object and a tag saying if the song is to be displayed "bold" or not
(this is probably going to be removed at some point).
The view now formats the song when it gets rendered on the screen, which
should speed-up quite a bit *loading* the song, at the expense of being
slower when the song are *displayed*. This will probably be speed up in
another commit, by introducing some kind of "formatting cache" to improve
the speed of formatting in a specific way a song, several times in a row.