current: remove manual storage & compatutation of the "boldness" of a row
Sonata used to store the "boldness" status (ie. the fact that a song with
a bold font is supposed to be the current one) directly inside the view's
model.
This had several drawbacks:

* it requires to be careful to correctly update this field when songs where
  changing, which meant removing the boldness from the previous song and
  setting it for the new one
* some part of code were a bit dodgy, by supposing that the last column of
  a model was actually storing this "bold" status
* it was some kind of conceptual violation of the Model/View principle

This is now IMO much clearer, cleaner and should also fix some bug related
to this "bold" feature as well.