formatting: formatting cache + usage in the Current list view
Introduce a new interface for the formatting module with some caching
capabilities. A small test with the previous interface and the new one
with formatting 10000+ songs and about 7 columns goes from about 20 seconds
to 3 seconds here.
The Current view now uses this cache and is also faster.

Sonata uses a bit more memory though, but the cache should be automatically
cleaned when the playlist changes.